+++ DTCD-HeaderPanel/src/Plugin.js
@@ -15,6 +15,9 @@ import {
 export class Plugin extends AppPanelPlugin {
   #vue;
   #workspaceSystem;
+  #interactionSystem;
+  #idAuthorizedUser = null;
+  static userEndpoint = '/dtcd_utils/v1/user?photo_quality=low';
 
   #settings = {
     showPageTitle: false,
@@ -33,7 +36,7 @@ export class Plugin extends AppPanelPlugin {
   constructor(guid, selector) {
     super(guid, selector);
     const eventSystem = new EventSystemAdapter('0.4.0', guid);
-    const interactionSystem = new InteractionSystemAdapter('0.4.0');
+    this.#interactionSystem = new InteractionSystemAdapter('0.4.0');
     this.#workspaceSystem = new WorkspaceSystemAdapter('0.4.0');
     const styleSystem = new StyleSystemAdapter('0.4.0');
     const router = new RouteSystemAdapter('0.1.0');
@@ -56,7 +59,7 @@ export class Plugin extends AppPanelPlugin {
     const data = {
       plugin: this,
       guid,
-      interactionSystem,
+      interactionSystem: this.#interactionSystem,
       eventSystem,
       workspaceSystem: this.#workspaceSystem,
       styleSystem,
@@ -77,6 +80,8 @@ export class Plugin extends AppPanelPlugin {
         },
       }
     }).$mount(selector);
+
+    this.#setSettingsFromLS();
   }
 
   onNewNotify(obj) {
@@ -107,6 +112,7 @@ export class Plugin extends AppPanelPlugin {
 
   setFormSettings(config) {
     this.setPluginConfig(config);
+    this.#saveConfigToLS();
   }
 
   getFormSettings() {
@@ -127,6 +133,47 @@ export class Plugin extends AppPanelPlugin {
           ],
         },
       ]
+    }
+  }
+
+  #getIdAuthorizedUser = async () => {
+    try {
+      const response = await this.#interactionSystem.GETRequest(Plugin.userEndpoint);
+      if (response?.data) {
+        const { id } = response.data;
+        this.#idAuthorizedUser = id;
+      }
+    } catch (error) {
+      this.#idAuthorizedUser = null;
+      throw error;
+    }
+    return;
+  }
+
+  #saveConfigToLS = async () => {
+    if (this.#idAuthorizedUser == null) await this.#getIdAuthorizedUser();
+    if (this.#idAuthorizedUser == null) return;
+
+    // not all settings from object '#settings' need to be saved to localStorage
+    const settingsToSave = {
+      notificationPosition: this.#settings.notificationPosition,
+    };
+
+    window.localStorage.setItem(
+      `${this.#idAuthorizedUser}:headerPanelConfig`,
+      JSON.stringify(settingsToSave)
+    );
+  }
+
+  #setSettingsFromLS = async () => {
+    if (this.#idAuthorizedUser == null) await this.#getIdAuthorizedUser();
+    if (this.#idAuthorizedUser == null) return;
+
+    const jsonSettings = window.localStorage.getItem(
+      `${this.#idAuthorizedUser}:headerPanelConfig`
+    );
+    if (jsonSettings) {
+      this.setPluginConfig(JSON.parse(jsonSettings));
     }
   }
 }